工单处理逻辑修改，积分相关接口文档及代码编写，增加部分页面js验证，优化部分页面显示效果

+++ src/com/hydom/credit/service/TrophyRecordService.java
@@ -30,7 +30,7 @@ public interface TrophyRecordService extends DAO<TrophyRecord> {
 	 * @param uid
 	 * @return
 	 */
-	public int countMonth(long uid);
+	public long countMonth(long uid);
 
 	/**
 	 * 统计指定用户兑换总次数
@@ -38,6 +38,6 @@ public interface TrophyRecordService extends DAO<TrophyRecord> {
 	 * @param uid
 	 * @return
 	 */
-	public int countAll(long uid);
+	public long countAll(long uid);
 
 }
